+++ packages/riverpod/lib/src/stream_provider/base.dart
@@ -123,6 +123,20 @@ class StreamProviderElement<T> extends ProviderElementBase<AsyncValue<T>>
   set state(AsyncValue<T> state) => setState(state);
 
   @override
+  void setState(AsyncValue<T> newState) {
+    AsyncValue<T> stateWithPrevious;
+
+    final previous = getState()?.requireState;
+    if (previous != null) {
+      stateWithPrevious = newState.copyWithPrevious(previous);
+    } else {
+      stateWithPrevious = newState;
+    }
+
+    super.setState(stateWithPrevious);
+  }
+
+  @override
   void create() {
     _streamNotifier.result ??= Result.data(_streamController.stream);